+++ garage/q_functions/__init__.py
@@ -1,2 +1,1 @@
 from garage.q_functions.base import QFunction
-from garage.q_functions.continuous_mlp_q_function import ContinuousMLPQFunction
