+++ garage/theano/q_functions/__init__.py
@@ -1,0 +1,2 @@
+from garage.theano.q_functions.continuous_mlp_q_function import \
+        ContinuousMLPQFunction
